refactor: move getOrderedRevisionsForEntityPage to helper/revision; chore: revision notes must be sorted; chore: use PagerElement for Entity-Revision-Table fix: fix a wrongly resolved commit in previous PR
